refactor: Renamed utils/test to utils/testing
